# app/web/swagger.yaml
swagger: "2.0"
info:
  description: "This is a work log for tracking different types of days for a user"
  version: "1.0.0"
  title: "Work Log"
  contact:
    email: "maxshapiro32@ibm.com"
  license:
    name: "Apache 2.0"
    url: "https://github.com/IBM/worklog/blob/master/LICENSE"
host: "localhost:5000"
basePath: "/api/v1"
tags:
- name: "Login"
  description: "Logging into a user account"
- name: "Users"
  description: "Operations associated with users"
schemes:
- "https"
paths:
  /login:
    post:
      tags:
      - "Login"
      summary: "Login to user account"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "data"
        description: "User credentials for logging into account"
        required: True
        schema:
          $ref: "#/definitions/Credentials"
      responses:
        200:
          description: "Successfully logged in"
        400:
          description: "Did not log in"
  /user/create:
    put:
      tags:
      - "Users"
      summary: "Create user"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "data"
        description: "User credentials for creating an account"
        required: True
        schema:
          $ref: "#/definitions/Credentials"
      responses:
        200:
          description: "Successfully created user account"
        400:
          description: "User already exists or user not created"
  /user/{username}/reset:
    put:
      tags:
      - "Users"
      summary: "Reset a user's password"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "username"
        required: True
        type: "string"
        description: "The username associated with a user's account"
      - in: "body"
        name: "data"
        description: "User credentials along with new password"
        required: True
        schema:
          $ref: "#/definitions/ResetCredentials"
      responses:
        200:
          description: "Successfully reset user's password"
        403:
          description: "Did not reset user's password"
  /user/{username}/settings:
    get:
      tags:
      - "Users"
      summary: "Gets a user's settings"
      description: ""
      produces:
      - "application/json"
      responses:
        200:
          description: "Successfully updated user's settings"
        403:
          description: "Not logged in"
        404:
          description: "No settings found"
    put:
      tags:
      - "Users"
      summary: "Updates a user's settings"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "data"
        description: "Updated user settings. Requires at least one setting to be provided"
        required: True
        schema:
          $ref: "#/definitions/Settings"
      responses:
        200:
          description: "Successfully updated user's settings"
        400:
          description: "No settings specified"
        403:
          description: "Not logged in"
        404:
          description: "No settings found"
  /user/{username}:
    get:
      tags:
      - "Users"
      summary: "Get work log data about a user"
      description: ""
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "username"
        required: True
        description: "The username associated with a user's account"
        type: "string"
      - in: "query"
        name: "date"
        description: "Date to get associated work log data. Follows the format 'YYYY-MM-DD'"
        type: "string"
      responses:
        200:
          description: "Successfully got requested work log data"
        400:
          description: "Invalid parameters given"
        403:
          description: "Not logged in"
        404:
          description: "No work log data found"
    post:
      tags:
      - "Users"
      summary: "Add work log data for a user"
      description: ""
      operationId: "handleUserData"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "username"
        required: True
        description: "The username associated with a user's account"
        type: "string"
      - in: "query"
        name: "date"
        required: True
        description: "Date to add work log data for. Follows the format 'YYYY-MM-DD'"
        type: "string"
      - in: "query"
        name: "type"
        required: True
        description: "The specific type of day data"
        type: "string"
        enum:
        - "office"
        - "remote"
        - "vacation"
        - "holiday"
        - "sick"
      - in: "query"
        name: "location"
        description: "Location of remote work. Only required when type=remote"
        type: "string"
      - in: "body"
        name: "notes"
        description: "Optional notes to add for an entry"
        schema:
          $ref: "#/definitions/Notes"
      responses:
        200:
          description: "Successfully added work log data"
        400:
          description: "Invalid parameters given"
        403:
          description: "Not logged in"
    put:
      tags:
      - "Users"
      summary: "Update work log data for a user"
      description: ""
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "username"
        required: True
        description: "The username associated with a user's account"
        type: "string"
      - in: "query"
        name: "date"
        required: True
        description: "Date to update associated work log data. Follows the format 'YYYY-MM-DD'"
        type: "string"
      - in: "query"
        name: "type"
        required: True
        description: "The specific type of day data"
        type: "string"
        enum:
        - "office"
        - "remote"
        - "vacation"
        - "holiday"
        - "sick"
      - in: "query"
        name: "location"
        description: "Location of remote work. Only required when type=remote"
        type: "string"
      - in: "body"
        name: "notes"
        description: "Optional notes to add for an entry"
        schema:
          $ref: "#/definitions/Notes"
      responses:
        200:
          description: "Successfully updated work log data"
        400:
          description: "Invalid parameters given"
        403:
          description: "Not logged in"
        404:
          description: "No work log data found"
    delete:
      tags:
      - "Users"
      summary: "Delete a user or work log data for a user"
      description: ""
      produces:
      - "application/json"
      parameters:
      - in: "path"
        name: "username"
        required: True
        description: "The username associated with a user's account"
        type: "string"
      - in: "query"
        name: "date"
        description: "Date to delete associated work log data. Follows the format 'YYYY-MM-DD'"
        type: "string"
      - in: "query"
        name: "year"
        description: "Year to delete associated work log data. Requires deleteyear=true"
        type: "integer"
      - in: "query"
        name: "deleteuser"
        description: "Boolean to determine if the user should be deleted"
        type: "boolean"
      - in: "query"
        name: "deleteall"
        description: "Boolean to determine if the all the user's work log data should be deleted"
        type: "boolean"
      - in: "query"
        name: "deleteyear"
        description: "Boolean to determine if the user's work log year data should be deleted. Requires year query parameter"
        type: "boolean"
      responses:
        200:
          description: "Successfully deleted work log data or user"
        400:
          description: "Invalid parameters given"
        403:
          description: "Not logged in"
        404:
          description: "No work log data or user to delete"
definitions:
  Credentials:
    type: "object"
    properties:
      username:
        type: "string"
      password:
        type: "string"
  ResetCredentials:
    type: "object"
    properties:
      username:
        type: "string"
      password:
        type: "string"
      new_password:
        type: "string"
  Notes:
    type: "object"
    properties:
      notes:
        type: "string"
  Settings:
    type: "object"
    properties:
      slack:
        type: "string"
      remote:
        type: "integer"
      vacation:
        type: "integer"
      holiday:
        type: "integer"
      sick:
        type: "integer"